feat: 펀딩 참여 기능 구현

+++ src/main/java/com/lovecloud/fundingmanagement/query/response/FundingListResponse.java
@@ -7,8 +7,8 @@ import java.util.List;
 public record FundingListResponse(
         Long fundingId,
         String title,
-        int targetAmount,
-        int currentAmount,
+        long targetAmount,
+        long currentAmount,
         FundingStatus status,
         LocalDateTime endDate,
         ProductOptionsSummary productOptions,
